[DDW-321] Moves types in respective files. Deletes redundant type definitions.
