+++ source/renderer/app/api/accounts/types.js
@@ -1,0 +1,9 @@
+export type AdaAccount = {
+  amount: number,
+  addresses: AdaAddresses,
+  name: string,
+  walletId: string,
+  index: number
+};
+
+export type AdaAccounts = Array<AdaAccount>;
