+++ source/renderer/app/api/addresses/types.js
@@ -1,0 +1,7 @@
+export type AdaAddress = {
+  id: string,
+  used: boolean,
+  changeAddress: boolean
+};
+
+export type AdaAddresses = Array<AdaAddress>;
